CSCQVAIN-233: Organize the Language drop down. (#62)
There was a bug via custom function which messed up the sorting which the backend has provided to us. In the case that we need to do sorting, we should customize our query to the backend instead to get valid results.

Also there was a query format which was causing issues.