CSCQVAIN-262: Idenfifiers has now example. (#75)
There was the example field in the schema, this is now in use for the String field. For this use the uiExample was added to base.vue, which means that when there is a need to have the example data visible for other components, that can be used for that specific purpose.