CSCQVAIN-178: user redirection on login. (#76)
The user ends up now to datasets view, which has the new dataset button on top right, thus we reduced one click from all the service users when they login. We can revert this change if needed in the future, if we get some new features for the welcome page, which the users would find useful.